Use `init`/`deinit` to manage Cowsay state
At this point, your `Cowsay` struct holds an allocator, a writer, and
even several "private" fields. This is no longer a simple struct that
can transparently have any of its fields modified at any time without
consequence; as such, it is best to provide an `init` function that is
guaranteed to create a `Cowsay` struct in a valid state, and then expect
the user to only interface with the struct through provided functions
rather than modifying fields directly at will.

Once we have an `init` function and an allocator within the `Cowsay`
struct, we may as well simplify/optimize our `cow` field to be allocated
with the user provided allocator rather than be backed by a buffer!

In fact, having such allocated fields within the struct is far more in
line with what users expect when they must past in an allocator to a
struct initialization.