Make local state, local
It's important (outside of just Zig, but in general programming) to
identify *local state* and try to minimize such local state bleeding
outside of its scope.

Local state does not refer only to spatial locality (where it is located
in the code), but also temporal locality! In this case, we have some
state -- such as maximum line width and offset -- that exist only per
provided message. There is no need to persist such state after a message
is printed, as these values no longer hold any meaning without the
message as context.

Therefore, we should try to avoid storing these values as fields of the
`Cowsay` struct, and instead initialize and use these values in their
relevant functions!